requirements.txt
aiofiles==23.1.0
aiogram==2.14.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
flake8==6.1.0
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.4
pycodestyle==2.11.1
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
typing_extensions==4.7.1
urllib3==2.1.0
yarl==1.9.3
